% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check-input.R
\name{check_date}
\alias{check_date}
\title{Check if date input is ISO-8601 formatted: YYYY-MM-DD}
\usage{
check_date(date)
}
\arguments{
\item{date}{A date that is ISO-8601 formatted: YYYY-MM-DD.}
}
\description{
Check if date input is ISO-8601 formatted: YYYY-MM-DD
}
\examples{
check_date("2015-5-15")
check_date("42139")
}

