% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check-input.R
\name{check_location}
\alias{check_location}
\title{Check if coordinate input is decimall formatted number}
\usage{
check_location(coordinate)
}
\arguments{
\item{coordinate}{A decimal-formatted number.}
}
\description{
Check if coordinate input is decimall formatted number
}
\examples{
check_location(latitude)
}

