% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check-input.R
\name{check_query}
\alias{check_query}
\title{Check if query string is created by \code{create_query()} function}
\usage{
check_query(strquery)
}
\arguments{
\item{strquery}{The structured query string created by \code{create_query()}
function}
}
\description{
Check if query string is created by \code{create_query()} function
}
\examples{
check_query(query)
}

