% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check-input.R
\name{check_token}
\alias{check_token}
\title{Check if API token exists or valid}
\usage{
check_token(token)
}
\arguments{
\item{token}{The API token from \code{get_token()} function.}
}
\description{
Check if API token exists or valid
}
\examples{
check_token(token)
}

