% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create-query.R
\name{create_query}
\alias{create_query}
\title{Create a weather request query string}
\usage{
create_query(latitude, longitude, x = NULL, y = NULL, date = NULL,
  attribute = NULL, gdd_method = NULL)
}
\arguments{
\item{latitude}{A decimal-formmated number indicate the latitude of a
location}

\item{longitude}{A decimal-formmated number indicate the longitude of a
location}

\item{x}{Optional, number of grids on x axis (latitude)}

\item{y}{Optional, number of grids on y axis (longitude)}

\item{date}{Optional, a matrix with start date, end date and plant date
generated by \code{\link{set_date}}. Without given start date, the package
will use \code{Sys.Date()} as a start date.}

\item{attribute}{Optional, a vector with weather attributes. Generated by
\code{\link{get_attribute}}. By default, it will use minTemperature,
maxTemperature and precip.}

\item{gdd_method}{Optional,}
}
\value{
A query object if succesful, otherwise \code{failure}
}
\description{
Create a query string with multiple options
}
\examples{
create_query(latitude, longitude)
create_query(latitude, longitude, x = 5, y = 5)
create_query(latitude, longitude, x = 5, y = 5, set_date(start_date, end_date))
create_query(latitude, longitude, date = set_date(start_date, plant_date =
plant_date), attribute = get_attribute(c("^acc", "gdd$")), gdd_method =
set_gdd(gdd_method = "modifiedstandard"))
}

