% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weatherawhere.R
\docType{package}
\name{httr}
\alias{httr}
\alias{httr-package}
\title{\pkg{weatherawhere} is a package for gathering accurate & agriculture-specific weather data.}
\description{
\code{weatherawhere} is organised around the two most common API functions to
allow access to aWhere Weather Terrain™ data R API:
\code{\link{get_token}}, \code{\link{create_query}},
\code{\link{send_query}}.
}
\details{
Requests can be modified by various config options like
\code{\link{get_attribute}}, \code{\link{set_date}},
\code{\link{set_gdd}}, \code{\link{set_grid}},
and \code{\link{set_gdd}}.
}

